% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genes.R
\name{annotate_genes}
\alias{annotate_genes}
\title{Order and annotate genes}
\usage{
annotate_genes(mat1)
}
\arguments{
\item{mat1}{RNA o prot matrix with unordered genes}
}
\value{
A list with two fields: 1. data the input matrix with only annotated genes
ordered by chromosome position 2. annot list with the corresponding
chromosome annotation. The number of data rows and the annot list have the same length
}
\description{
Function to get annotation and order of genes, returning the input data sorted
and the chromosome annotation.
}
\examples{
mat1 <- annotate_genes(d$mat1)
}
