% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mirnas.R
\name{annotate_mirnas}
\alias{annotate_mirnas}
\title{Order and annotate miRNAs}
\usage{
annotate_mirnas(mat3)
}
\arguments{
\item{mat3}{miRNA matrix with unordered miRNAs}
}
\value{
A list with two fields: 1. data the input matrix with only annotated miRNAs
ordered by chromosome position 2. annot list with the corresponding
chromosome annotation. The number of data rows and the annot list have the same length
}
\description{
Function to get annotation and order of miRNAs, returning the input data sorted
and the chromosome annotation.
}
\examples{
mat3 <- annotate_genes(d$mat3)
}
