% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pc_values.R
\name{compute_pc_values}
\alias{compute_pc_values}
\title{From list to dataframe}
\usage{
compute_pc_values(mat1, mat2, mat3, corr_method, out_folder, ncores)
}
\arguments{
\item{mat1}{mRNA matrix}

\item{mat2}{prot matrix}

\item{mat3}{miRNA matrix}

\item{corr_method}{Correlation method for partial correlation. The allowed methods are "pearson", "kendall", "spearman"}

\item{out_folder}{Name of the output folder in which results will be saved}

\item{ncores}{Number of cores to be used to parallelize the code. Suggested numers are 1 for Windows users, from 1 to max number of cores for Linux users.}
}
\value{
A list with two fields: 1. estimate estimate value
2. p.value pvalue. The two lists have the same length
}
\description{
From list data structure to ESTIMATES data.frame and PVALUE dataframe
}
\examples{
a <- compute_pc_values(mat1, mat2, mat3, corr_method, out_folder, ncores)
}
