% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_corr.R
\name{compute_standard_corr}
\alias{compute_standard_corr}
\title{Function to compute standard correlation between two matrices}
\usage{
compute_standard_corr(mat1, mat2, mat1_name, mat2_name)
}
\arguments{
\item{mat1}{first matrix, genes on rows, samples on columns}

\item{mat2}{second matrix, genes on rows, samples on columns. The number of
samples must be the same}

\item{mat1_name}{name of the first matrix, e.g., "RNA"}

\item{mat2_name}{name of the second matrix, e.g., "miRNA"}
}
\value{
A list with three fields: 1. cormatrix: correlation matrix estimates
2. pmatrix: pvalue matrix, without fdr correction  3. table: dataframe with
all previous info in a table
}
\description{
Function to compute standard correlation between two matrices
}
\examples{
d <- compute_standard_corr(mat1, mat2, mat1_name, mat2_name)

}
