% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_fu_single.R
\name{cor_fu_single}
\alias{cor_fu_single}
\title{Correlation between one gene and one miRNA}
\usage{
cor_fu_single(mi, gene, m, mat1, mat2, mat3)
}
\arguments{
\item{mi}{miRNA name}

\item{gene}{gene name (COMMON gene notation)}

\item{m}{correlation method ("pearson", "kendall", "spearman")}

\item{mat1}{mRNA matrix}

\item{mat2}{miRNA matrix}
}
\value{
A list with two fields: 1. estimate estimate value
}
\description{
Function to get the correlation between one gene and one mirna
}
\examples{
a <- cor_fu_single(mi, gene, m, mat1, mat2, mat3)
2. p.value pvalue.
}
