% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_func_vect.R
\name{corr_func_vect}
\alias{corr_func_vect}
\title{Correlation between one gene and all miRNA}
\usage{
corr_func_vect(gene, m, mat1, mat2, mat3)
}
\arguments{
\item{gene}{gene name (COMMON gene notation)}

\item{m}{corralation method ("pearson", "kendall", "spearman")}

\item{mat1}{mRNA matrix}

\item{mat2}{miRNA matrix}
}
\value{
A list with two fields: 1. estimate estimate value
2. p.value pvalue.
}
\description{
Function to get the correlation between one gene and all mirna
}
\examples{
a <- corr_func_vect(gene, condv, m, mat1, mat2, mat3)

}
