% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relevant_improvement.R
\name{est_minus_single}
\alias{est_minus_single}
\title{Function to compute improvement}
\usage{
est_minus_single(gene, pval, single_corr_df)
}
\arguments{
\item{gene}{Common gene name}

\item{single_corr_df}{single correlation values}

\item{pvalue}{Gene name pvalue}
}
\value{
P value improvement
}
\description{
Function to compute improvement
}
