% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relevant_improvement.R
\name{fix_names}
\alias{fix_names}
\title{Function to fix mirna names}
\usage{
fix_names(name)
}
\arguments{
\item{name}{miRNA name}
}
\value{
Fixed miRNA name
}
\description{
put miRNA name in the standard format
}
