% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_corrs_to_df.R
\name{from_corrs_to_df}
\alias{from_corrs_to_df}
\title{This function transforms corrs into ESTIMATES and PVALUE DATAFRAMES}
\usage{
from_corrs_to_df(corrs, genes_proteins, condv)
}
\arguments{
\item{corrs}{A list with two fields: 1. estimate estimate value
2. p.value pvalue.}

\item{genes_proteins}{List of gene/proteins}

\item{condv}{List of miRNAs}
}
\value{
A list with two fields: 1. estimate estimate value
2. p.value pvalue. The two lists have the same length
}
\description{
From list data structure to ESTIMATES data.frame and PVALUE dataframe
}
