% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relevant_improvement.R
\name{get_relevant_improvement}
\alias{get_relevant_improvement}
\title{Function to get matrix with only relevant improvements}
\usage{
get_relevant_improvement(mat1, mat2, pval, corr_method, diff_th)
}
\arguments{
\item{mat1}{mRNA matrix}

\item{mat2}{prot matrix}

\item{pval}{pval partial correlation  matrix}

\item{corr_method}{"pearson", "spearman", "kendall"}

\item{diff_th}{improvement threashold}
}
\value{
matrix with only relevant improvements (differences between partial correlation pvalue and
standard correlation pvalue)
}
\description{
Function to get matrix with only relevant improvements
}
\examples{
mat <- get_relevant_improvement(mat1, mat2, pval, corr_method, diff_th)
}
