% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_test.R
\name{hyper_test}
\alias{hyper_test}
\title{Computing hypergeometric test}
\usage{
hyper_test(to_validate, universe, mirDB, ts, tarbase, outfile)
}
\arguments{
\item{to_validate}{dataframe of elements to be validated.}

\item{universe}{all mirnas_gene couples potentially validated.}

\item{mirDB}{list of mirDB validated elements}

\item{ts}{list of TargetScan validated elements}

\item{tarbase}{list of mirTarBase validated elements}

\item{outfile}{filename with the path of the folder which contains the output}
}
\description{
Function to compute te hyper geometric test
}
\examples{
 hyper_test(to_validate, universe, mirDB, ts, tarbase, out_folder)

}
