% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalist_preproc.R
\name{minimalist_preproc}
\alias{minimalist_preproc}
\title{Minimalist preprocessing to MoPc}
\usage{
minimalist_preproc(mat1, mat2, mat3)
}
\arguments{
\item{mat1}{The first matrix. Usually mRNA matrix with samples on the columns and genes on the rows.
The gene names must be in the "common name" format (e.g., ABL1)}

\item{mat2}{The second matrix. Usually prot matrix with samples on the columns and proteins on the rows.
The protein names must be in the "common name" format (e.g., ABL1)}

\item{mat3}{The third matrix. Usually miRNA matrix with samples on the columns and miRNAs on the rows.
The miRNA names must be in the "hsa-miR-XXXXXX" format (e.g., hsa-miR-200c-3p)}
}
\value{
It quits the execution if there are not enough samples or genes to proceed.
Otherwise, it returns a list with the three matrices, properly subset to common genes and common samples.
}
\description{
Minimalist preprocessing to MoPc
}
\examples{
d <- minimalist_preproc(rna, prot, mirna)

}
