% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relevant_improvement.R
\name{normal_cor}
\alias{normal_cor}
\title{Function to compute standard correlation}
\usage{
normal_cor(gene, mat1, mat2, corr_method)
}
\arguments{
\item{gene}{Common gene name}

\item{mat1}{First matrix for std corr}

\item{mat2}{Second matrix for std corr}

\item{corr_method}{One of "pearson", "spearman", "kendall"}
}
\value{
A list with 2 fields: estimate and pvalue
}
\description{
Function to compute standard correlation
}
