% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_corr_heatmap_plot.R
\name{partial_corr_heatmap_plot}
\alias{partial_corr_heatmap_plot}
\title{Function to plot standard correlation between two matrices in a pdf file}
\usage{
partial_corr_heatmap_plot(
  to_validate,
  genes_proteins,
  condv,
  gene_annot,
  condv_annot,
  mat,
  out_folder
)
}
\arguments{
\item{to_validate}{dataframe to plot eventually with NAs}

\item{genes_proteins}{list of all genes}

\item{condv}{list of all mirnas}

\item{gene_annot}{List containing the chromosome location of each cor_matrix
column}

\item{condv_annot}{List containing the chromosome location of each cor_matrix
rows}

\item{mat}{matrix}

\item{out_folder}{Path to the desired output folder}
}
\description{
Function to plot standard correlation between two matrices in a pdf file
}
