% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_corr.R
\name{plot_heatmap_corr}
\alias{plot_heatmap_corr}
\title{Function to plot standard correlation between two matrices in a pdf file}
\usage{
plot_heatmap_corr(
  cor_matrix,
  column_annot,
  row_annot,
  filename,
  row_name,
  column_name
)
}
\arguments{
\item{cor_matrix}{correlation matrix to plot eventually with NAs}

\item{column_annot}{List containing the chromosome location of each cor_matrix
column}

\item{row_annot}{List containing the chromosome location of each cor_matrix
rows}

\item{filename}{Path and name where the plot will be saved (e.g.,
"/results/Heatmap")}

\item{row_name}{name to be put in the plot for rows, e.g., "mRNA"}

\item{column_name}{name to be put in the plot for columns, e.g., "miRNA"}
}
\description{
Function to plot standard correlation between two matrices in a pdf file
}
