% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prot}
\alias{prot}
\title{Example data for the pipeline}
\format{
A list withthe elements
}
\source{
{Created as an example}
}
\usage{
prot
}
\description{
Example data sampled from breast cancer dataset
}
\examples{
data(rna)
data(prot)
data(mirna)

}
\keyword{datasets}
