% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_info.R
\name{validation_info}
\alias{validation_info}
\title{Getting validation info}
\usage{
validation_info(mat, condv, genes_proteins, ncores, from_cor_mat = TRUE)
}
\arguments{
\item{mat}{the correlation matrix (eventually with NAs).
Genes on rows - mirnas on columns. Also, it can be a list of
significant genes-mirnas with pvalue (to_validate dataframe).}

\item{condv}{all mirnas in correlation matrix}

\item{genes_proteins}{all genes/proteins in correlation matrix}

\item{ncores}{max number of cores to prallelize the process}

\item{from_cor_mat}{default to TRUE, set FALSE if you input the to_validate
dataframe}
}
\value{
A list with four fields: list(a=to_validate, b=universe, c =mirDB, d=ts, e=tarbase)
}
\description{
Function to know if a gene-mirna couple is validated in mirDB, TargetScan,
mirtarBase
}
